requirements.txt
streamlit==1.29.0
pymongo==4.6.1
ultralytics==8.1.0
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
pandas==2.1.4
numpy==1.24.4
